feat(tys/xmp): add idea of "XmpElement"
the difference between this and `XmpValue` is that a Value doesn't
contain info about where it came from.

on the other hand, `XmpElement` says how it was derived from the XML.

in the future, this'll make serialization muuuuch better. but, for now,
it just helps with debugging a lot ;D