fix(tys/xmp/XmpValue::Alternatives): add tag info
...onto each variant.

without this info, we wouldn't be able to reconstruct a list of
alternatives. which would suck!

note that we'll need additional changes for proper re-serialization in
the future, such as maintaining internal information about what info was
already available in the attributes and such. (it's pretty mean to get
rid of another program's custom functionality lol)